// src/lib/enum.ts
export enum TEAM_TARGET {
  PLAN_MAINTENANCE = 'PlanMaintenance'
}
export enum leaseAgreementStageEnum {
  New = 1,
  Confirmed = 2,
  Terminate = 3,
  EarlyTerminate = 4,
  Closed = 5
}

export enum FeeTypeEnum {
  rentFee = 1,
  managerFee = 2
}
export enum distributionChannelEnum {
  local = '10',
  other = '20'
}
export enum MaterialSAPEnum {
  MON_Lease = 'ASD00000107',
  QR_Lease = 'ASD00000108',
  YR_Lease = 'ASD00000134',
  MON_ManagerFee = 'ASD00000109',
  QR_ManagerFee = 'ASD00000110',
  YR_ManagerFee = 'ASD00000111'
}
export enum countryOptionEnum {
  Vietnam = 232
}

export enum paymentScheduleStatusEnum {
  unpaid = 1,
  paid = 2,
  cancel = 3
}